Extract GapfillDeltaTransition into separate module
